backend/ecoalert-backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
supabase==2.4.0
firebase-admin==6.4.0
python-dotenv==1.0.1
Pillow==10.2.0
